requirements.txt
pystray
sv_ttk
notify-py
pillow
pyautogui
